pyproject.toml
[project]
name = "sat-water-temps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-xray-sdk>=2.14.0",
    "boto3>=1.35.0",
    "earthaccess>=0.12.0",
    "cmap>=0.4.0",
    "numpy>=2.0.2",
    "pillow>=11.0.0",
    "pandas>=2.3.3",
    "pyarrow>=19.0.0",
    "pystac-client>=0.8.3",
    "python-dotenv>=1.1.1",
    "rasterio>=1.4.3",
    "requests>=2.32.5",
    "shapely>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
